use int instead of double

+++ r6adssensitivitycalculator/app/src/main/java/com/poorskill/r6adssensitivitycalculator/utility/SensitivityCalculator.java
@@ -4,8 +4,8 @@ public final class SensitivityCalculator {
     private static final double[] fovMultiplier = {.9, .59, .49, .42, .35, .3, .22, .092};
     private static final double[] adsMultiplier = {.6, .59, .49, .42, .35, .3, .22, .14};
 
-    public static double[] calculateNewAdsSensitivity(int oldAds, int fov, double aspectRatio) {
-        double[] result = new double[8];
+    public static int[] calculateNewAdsSensitivity(int oldAds, int fov, double aspectRatio) {
+        int[] result = new int[8];
         double horizontalFOV = calculateHorizontalFOV(fov, aspectRatio);
         double verticalFOV = horizontalFOV > 150 ? calculateVerticalFOV(aspectRatio) : fov;
 
@@ -19,8 +19,8 @@ public final class SensitivityCalculator {
         return Math.tan(Math.toRadians(fovMultiplier * verticalFOV / 2.)) / Math.tan(Math.toRadians(verticalFOV / 2.));
     }
 
-    private static double calculateNewAds(double adsMultiplier, double fovAdjustment, double oldAds) {
-        return (adsMultiplier / fovAdjustment) * oldAds;
+    private static int calculateNewAds(double adsMultiplier, double fovAdjustment, double oldAds) {
+        return (int) ((adsMultiplier / fovAdjustment) * oldAds);
     }
 
     private static double calculateVerticalFOV(double aspectRatio) {
